requirements.txt
fastapi==0.111.0 
uvicorn==0.30.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9 
google-generativeai==0.6.0
PyPDF2==3.0.1
Pillow==10.3.0
pydantic-settings==2.3.4 